Refactor Code Base
- [+] chore(crew.go): remove unused import statement
- [+] refactor(crew.go): move performTaskWithRetries function to error_and_retry.go
- [+] feat(error_and_retry.go): add performTaskWithRetries function with retry logic and error handling
