+++ worker/error_and_retry.go
@@ -14,6 +14,71 @@ import (
 	"k8s.io/client-go/kubernetes"
 )
 
+// performTaskWithRetries tries to execute a task, with retries on failure.
+// It honors the cancellation signal from the context and ceases retry attempts
+// if the context is cancelled. If the task remains incomplete after all retries,
+// it returns an error detailing the failure.
+//
+// Parameters:
+//
+//	ctx context.Context: Context for task cancellation and timeouts.
+//	clientset *kubernetes.Clientset: Kubernetes API client for executing tasks.
+//	shipsNamespace string: Kubernetes namespace for task execution.
+//	task configuration.Task: Task to be executed.
+//	results chan<- string: Channel for reporting task execution results.
+//	workerIndex int: Index of the worker for contextual logging.
+//	taskStatus *TaskStatusMap: Map to track and control the status of tasks.
+//
+// Returns:
+//
+//	error: Error if the task fails after all retry attempts.
+func performTaskWithRetries(ctx context.Context, clientset *kubernetes.Clientset, shipsNamespace string, task configuration.Task, results chan<- string, workerIndex int, taskStatus *TaskStatusMap) error {
+	// Define the operation to be retried.
+	operation := func() (string, error) {
+		// Attempt to perform the task.
+		err := performTask(ctx, clientset, shipsNamespace, task, workerIndex)
+		return task.Name, err // Return the task name along with the error.
+	}
+
+	// Create a RetryPolicy instance with the task's retry settings.
+	retryPolicy := RetryPolicy{
+		MaxRetries: task.MaxRetries,
+		RetryDelay: task.RetryDelayDuration,
+	}
+
+	// Use the RetryPolicy's Execute method to perform the operation with retries.
+	err := retryPolicy.Execute(ctx, operation, func(message string, fields ...zap.Field) {
+		// This is a placeholder for the actual logging function.
+		// Replace this with the actual function to log retries.
+		// For example: navigator.LogInfoWithEmoji or navigator.LogErrorWithEmoji
+		// Combine emojis with a space for readability.
+		emojiField := fmt.Sprintf("%s %s", constant.ErrorEmoji, language.PirateEmoji)
+		navigator.LogErrorWithEmoji(emojiField, message, fields...)
+	})
+
+	if err != nil {
+		// Additional error handling logic
+		if apierrors.IsConflict(err) {
+			// Handle conflict-specific errors
+			conflictResolved := handleConflictError(ctx, clientset, shipsNamespace, &task)
+			if conflictResolved {
+				// Conflict resolved, retry the operation
+				return performTaskWithRetries(ctx, clientset, shipsNamespace, task, results, workerIndex, taskStatus)
+			}
+		} else {
+			// Handle generic errors that are not conflicts
+			handleGenericError(ctx, err, task.MaxRetries, &task, workerIndex, task.MaxRetries, task.RetryDelayDuration)
+		}
+
+		handleFailedTask(task, taskStatus, shipsNamespace, err, results, workerIndex)
+		return fmt.Errorf(language.ErrorFailedToCompleteTask, task.Name, task.MaxRetries)
+	}
+
+	// If the operation was successful, handle the success.
+	handleSuccessfulTask(task, results, workerIndex)
+	return nil
+}
+
 // logRetryAttempt logs a warning message indicating a task retry attempt with the current count.
 // It includes the task name and the error that prompted the retry. The maxRetries variable is used
 // to indicate the total number of allowed retries.
@@ -29,7 +94,7 @@ func logRetryAttempt(taskName string, attempt int, err error, maxRetries int, lo
 	message := fmt.Sprintf(language.ErrorDuringTaskAttempt, attempt+1, maxRetries, err)
 	fields := []zap.Field{
 		zap.String(tasK, taskName),
-		zap.Int(attempT, attempt+1),
+		zap.Int(attempT, attempt),
 		zap.Int(maXRetries, maxRetries),
 		zap.Error(err),
 	}
@@ -77,6 +142,9 @@ func logFinalError(shipsnamespace string, taskName string, err error, maxRetries
 // Returns:
 //
 //	shouldContinue bool: A boolean indicating whether the task should be retried or not.
+//
+// Deprecated: Already Sync with Retry Policy which is better for reduce complex and free resource channel for go routines (known as gopher).
+// so this function are not longer used.
 func handleTaskError(ctx context.Context, clientset *kubernetes.Clientset, shipsnamespace string, err error, attempt int, task *configuration.Task, workerIndex int, maxRetries int, retryDelay time.Duration) (shouldContinue bool) {
 	if ctx.Err() != nil {
 		return false
@@ -130,7 +198,7 @@ func handleConflictError(ctx context.Context, clientset *kubernetes.Clientset, s
 //
 //	bool: A boolean indicating whether the task should be retried or not.
 func handleGenericError(ctx context.Context, err error, attempt int, task *configuration.Task, workerIndex int, maxRetries int, retryDelay time.Duration) bool {
-	logRetryAttempt(task.Name, attempt, err, maxRetries, navigator.Logger.Error)
+	logRetryAttempt(task.Name, attempt, err, maxRetries, navigator.Logger.Info)
 
 	// Wait for the next attempt, respecting the context cancellation.
 	if !waitForNextAttempt(ctx, retryDelay) {
